Move ParsedMessage and parseMessage back to message.ts
